Remove legacy code from the backend
