--- backend/backend/ecg_app/management/commands/populate_dummy_data.py
@@ -1,58 +1,0 @@
-import random
-from django.core.management.base import BaseCommand
-from django.contrib.auth.models import User
-from ecg_app.models import (
-    Profile, Quiz, Question, Choice, QuizAttempt,
-    QuestionAttempt, UserStatistics, EcgSamples
-)
-
-class Command(BaseCommand):
-    help = "Populate the database with dummy values for all models"
-
-    def handle(self, *args, **kwargs):
-
-        # Create admin user
-        if not User.objects.filter(username='admin').exists():
-            User.objects.create_superuser('admin', 'admin@example.com', 'admin123')
-            self.stdout.write(self.style.SUCCESS('Admin user created successfully'))
-
-        # Add users
-        for i in range(2):
-            user, created = User.objects.get_or_create(
-                username=f"user{i+1}",
-                email=f"user{i+1}@example.com",
-                first_name=f"User {i+1}",
-                last_name=f"Last Name {i+1}",
-                is_staff=random.choice([True, False]),
-            )
-            if created:
-                user.set_password("password")
-                user.save()
-                Profile.objects.get_or_create(user=user)
-
-
-        # Add quiz attempts
-        for user in User.objects.all():
-            for quiz in Quiz.objects.all():
-                attempt = QuizAttempt.objects.create(user=user, quiz=quiz)
-                for question in quiz.questions.all():
-                    selected_choice = random.choice(question.choices.all())
-                    QuestionAttempt.objects.create(
-                        quiz_attempt=attempt,
-                        question=question,
-                        selected_choice=selected_choice,
-                        is_correct=selected_choice.is_correct,
-                    )
-
-        # Add user statistics
-        for user in User.objects.all():
-            UserStatistics.objects.get_or_create(
-                user=user,
-                total_quizzes_taken=QuizAttempt.objects.filter(user=user).count(),
-                total_correct_answers=sum(
-                    q.is_correct for q in QuestionAttempt.objects.filter(quiz_attempt__user=user)
-                ),
-                total_questions_answered=QuestionAttempt.objects.filter(quiz_attempt__user=user).count(),
-            )
-
-        self.stdout.write(self.style.SUCCESS("Dummy data populated successfully!"))
