+++ backend/backend/ecg_app/urls.py
@@ -15,7 +15,6 @@ router.register(r'ecg-snomed', views.EcgSnomedViewSet, basename='ecg-snomed')
 router.register(r'ecg-samples-doc-labels', views.EcgSamplesDocLabelsViewSet, basename='ecg-samples-doc-labels')
 router.register(r'ecg-samples-snomed', views.EcgSamplesSnomedViewSet, basename='ecg-samples-snomed')
 router.register(r'profiles', views.ProfileViewSet, basename='profiles')
-router.register(r'statistics', views.UserStatisticsViewSet, basename='statistics')
 router.register(r'quizzes', views.QuizViewSet, basename='quizzes')
 router.register(r'questions', views.QuestionViewSet, basename='questions')
 router.register(r'choices', views.ChoiceViewSet, basename='choices')
